optimize preference views
