+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/ui/devkit/components/settings/RightArrowIcon.kt
@@ -1,0 +1,14 @@
+package com.github.klee0kai.thekey.core.ui.devkit.components.settings
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+
+@Composable
+fun RightArrowIcon() {
+    Icon(
+        imageVector = Icons.AutoMirrored.Default.KeyboardArrowRight,
+        contentDescription = null,
+    )
+}
